requirements.txt
asgiref==3.2.10
backcall==0.2.0
captcha==0.3
click==7.1.2
colorama==0.4.3
decorator==4.4.2
Django==3.0.8
django-cors-headers==3.4.0
djangorestframework==3.11.0
Flask==1.1.2
ipython==7.16.1
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
MarkupSafe==1.1.1
parso==0.7.1
pickleshare==0.7.5
Pillow==7.2.0
prompt-toolkit==3.0.5
Pygments==2.6.1
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
traitlets==4.3.3
wcwidth==0.2.5
Werkzeug==1.0.1
